requirements.txt
Django
djangorestframework
django-rest-swagger